// aga_2/src/data/interfaces.ts
export interface ILogin_Herokuapp {
    title: string,
    subheader: string
    usernameLabel: string,
    passwordLabel: string,
    button: string
}

export interface ILogin_karpovich {
    title: string;
    usernameInput: string;
    passwordInput: string;
    submitButton: string;
    registerButton: string;
    errorMessage: string;
    successMessage: string;
    backButton: string
}
  
export interface IRegistration_karpovich {
    title: string;
    usernameInput: string;
    passwordInput: string;
    registerButton: string;
    backButton: string;
    errorMessage: string;
}